+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Add basic prometheus metrics for counting sessions, clients and logins ( in addition to default metrics )
+
 ## [4.4.0] [r22.11]
 ### Added
 - Add logging of actions of Admin Client API ( restart clients, turn off clients, modify time, etc )